requirements.txt
Jinja2==3.1.2
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
python-dotenv==0.20.0